% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_bcftools_trio_stats.R
\name{summarize_bcftools_trio_stats}
\alias{summarize_bcftools_trio_stats}
\title{summarize bcftools trio stats output}
\usage{
summarize_bcftools_trio_stats(d, alpha = 0.05, decimals = 3)
}
\arguments{
\item{d}{a data.frame (typically file generated with bcftools --trio-stats)}

\item{alpha}{confidence level for binary CIs}

\item{decimals}{how many decimals should SERs be rounded with?}
}
\description{
summarize bcftools trio stats output
}
